Extract heavy non-strict version
