+++ base.js
@@ -1,0 +1,18 @@
+'use strict';
+const ipRegex = require('ip-regex');
+
+module.exports = (opts, tlds) => {
+	opts = Object.assign({}, opts);
+
+	const protocol = `(?:(?:[a-z]+:)?//)${tlds ? '?' : ''}`;
+	const auth = '(?:\\S+(?::\\S*)?@)?';
+	const ip = ipRegex.v4().source;
+	const host = '(?:(?:[a-z\\u00a1-\\uffff0-9]-*)*[a-z\\u00a1-\\uffff0-9]+)';
+	const domain = '(?:\\.(?:[a-z\\u00a1-\\uffff0-9]-*)*[a-z\\u00a1-\\uffff0-9]+)*';
+	const tld = `(?:\\.(?:${tlds || '[a-z\\u00a1-\\uffff]{2,}'}))\\.?`;
+	const port = '(?::\\d{2,5})?';
+	const path = '(?:[/?#][^\\s"]*)?';
+	const regex = `(?:${protocol}|www\\.)${auth}(?:localhost|${ip}|${host}${domain}${tld})${port}${path}`;
+
+	return opts.exact ? new RegExp(`(?:^${regex}$)`, 'i') : new RegExp(regex, 'ig');
+};
