+++ index.js
@@ -1,19 +1,5 @@
-'use strict';
-const ipRegex = require('ip-regex');
-const tlds = require('tlds');
+const buildRegex = require('./base');
 
-module.exports = opts => {
-	opts = Object.assign({strict: true}, opts);
-
-	const protocol = `(?:(?:[a-z]+:)?//)${opts.strict ? '' : '?'}`;
-	const auth = '(?:\\S+(?::\\S*)?@)?';
-	const ip = ipRegex.v4().source;
-	const host = '(?:(?:[a-z\\u00a1-\\uffff0-9]-*)*[a-z\\u00a1-\\uffff0-9]+)';
-	const domain = '(?:\\.(?:[a-z\\u00a1-\\uffff0-9]-*)*[a-z\\u00a1-\\uffff0-9]+)*';
-	const tld = `(?:\\.${opts.strict ? '(?:[a-z\\u00a1-\\uffff]{2,})' : `(?:${tlds.join('|')})`})\\.?`;
-	const port = '(?::\\d{2,5})?';
-	const path = '(?:[/?#][^\\s"]*)?';
-	const regex = `(?:${protocol}|www\\.)${auth}(?:localhost|${ip}|${host}${domain}${tld})${port}${path}`;
-
-	return opts.exact ? new RegExp(`(?:^${regex}$)`, 'i') : new RegExp(regex, 'ig');
-};
+// This needs to be called this way because
+// buildRegex accepts a second parameter, while `url-regex` doesn't
+module.exports = opts => buildRegex(opts);
