+++ non-strict.js
@@ -1,0 +1,4 @@
+const tlds = require('tlds');
+const buildRegex = require('./base');
+
+module.exports = opts => buildRegex(opts, tlds.join('|'));
